handle postcss in render_static
