+++ crates/next-core/js/package.json
@@ -14,6 +14,7 @@
     "css.escape": "^1.5.1",
     "next": "^13.0.6",
     "platform": "1.3.6",
+    "postcss": "^8.4.20",
     "react-dom": "^18.2.0",
     "react": "^18.2.0",
     "source-map": "0.8.0-beta.0",
