+++ crates/turbopack-node/js/src/ipc/evaluate.ts
@@ -13,27 +13,31 @@ type IpcOutgoingMessage = {
 
 const ipc = IPC as Ipc<IpcIncomingMessage, IpcOutgoingMessage>;
 
-export const run = (getValue: () => any) => {
-  (async () => {
-    while (true) {
-      const msg = await ipc.recv();
+export const run = async (
+  getValue: (...deserializedArgs: any[]) => any,
+  ...args: string[]
+) => {
+  while (true) {
+    const msg = await ipc.recv();
 
-      switch (msg.type) {
-        case "evaluate": {
-          const value = await getValue();
-          await ipc.send({
-            type: "jsonValue",
-            data: JSON.stringify(value),
-          });
-          break;
-        }
-        default: {
-          console.error("unexpected message type", msg.type);
-          process.exit(1);
-        }
+    switch (msg.type) {
+      case "evaluate": {
+        const deserializedArgs = args.map((arg) => JSON.parse(arg));
+        const value = await getValue(...deserializedArgs).catch(
+          (err: Error) => {
+            return ipc.sendError(err);
+          }
+        );
+        await ipc.send({
+          type: "jsonValue",
+          data: JSON.stringify(value),
+        });
+        break;
+      }
+      default: {
+        console.error("unexpected message type", msg.type);
+        process.exit(1);
       }
     }
-  })().catch((err) => {
-    ipc.sendError(err);
-  });
+  }
 };
